spit in 3 profiles
