+++ src/main/java/eu/dissco/core/digitalspecimenprocessor/Profiles.java
@@ -6,6 +6,12 @@ public class Profiles {
 
 	public static final String WEB = "web";
 
+	public static final String SPECIMEN_RABBIT_MQ = "specimen-rabbitmq";
+
+	public static final String MEDIA_RABBIT_MQ = "media-rabbitmq";
+
+	public static final String ER_RABBIT_MQ = "er-rabbitmq";
+
 	private Profiles() {
 	}
