+++ src/main/java/eu/dissco/core/digitalspecimenprocessor/service/rabbitmqconsumer/SpecimenRabbitMqConsumerService.java
@@ -1,0 +1,36 @@
+package eu.dissco.core.digitalspecimenprocessor.service.rabbitmqconsumer;
+
+import eu.dissco.core.digitalspecimenprocessor.Profiles;
+import eu.dissco.core.digitalspecimenprocessor.domain.specimen.DigitalSpecimenEvent;
+import eu.dissco.core.digitalspecimenprocessor.service.preprocessing.SpecimenPreprocessingService;
+import java.util.List;
+import java.util.Objects;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.amqp.rabbit.annotation.RabbitListener;
+import org.springframework.context.annotation.Profile;
+import org.springframework.messaging.handler.annotation.Payload;
+import org.springframework.stereotype.Service;
+import tools.jackson.databind.ObjectMapper;
+
+@Service
+@Slf4j
+@Profile(Profiles.SPECIMEN_RABBIT_MQ)
+@RequiredArgsConstructor
+public class SpecimenRabbitMqConsumerService {
+
+	private final ObjectMapper mapper;
+
+	private final SpecimenPreprocessingService processingService;
+
+	@RabbitListener(queues = { "${rabbitmq.queue-name-specimen:digital-specimen-queue}" },
+			containerFactory = "consumerBatchContainerFactory")
+	public void getMessages(@Payload List<String> messages) {
+		var events = messages.stream()
+			.map(message -> mapper.readValue(message, DigitalSpecimenEvent.class))
+			.filter(Objects::nonNull)
+			.toList();
+		processingService.handleMessages(events);
+	}
+
+}
